% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-prepare.R
\name{prepare}
\alias{prepare}
\title{Prepare data to create a \code{REGObject}}
\usage{
prepare(data, vars_y, vars_x, vars_c)
}
\arguments{
\item{data}{A \code{data.table} storing input data.}

\item{vars_y}{Column names storing the \strong{predicted variables}.}

\item{vars_x}{Column names storing the \strong{focal variables}.}

\item{vars_c}{Column names storing the \strong{Covariables}.}
}
\description{
It is recommended that the input \code{data} has been cleaned for modeling purposes. While setting "factor(cyl)" is acceptable in the \verb{vars_*} options, ensuring data cleanliness will improve the overall readability.
}
\examples{
x <- prepare(
  data = mtcars,
  vars_y = "mpg",
  vars_x = c("factor(cyl)", colnames(mtcars)[3:5]),
  vars_c = c(colnames(mtcars)[8:9], "factor(gear)")
)
x
}
