% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{REGObject-class}
\alias{REGObject-class}
\title{REGObject: class representing a list of regression model}
\description{
Contains fields storing data and methods to build, process and visualize
a list of regression model.
Currently, this class is designed for CoxPH and GLM regression models.
}
\section{Slots}{

\describe{
\item{\code{data}}{A \code{data.table} storing input data.}

\item{\code{mdata}}{A \code{data.table} storing cleaned data for modeling.}

\item{\code{vars_x}}{Focal variables.}

\item{\code{var_y}}{Predicted variables or formulas}

\item{\code{vars_c}}{Covariables.}

\item{\code{args}}{Other arguments used for building regression models.}

\item{\code{models}}{Models built.}

\item{\code{results}}{Model results, a object of \code{parameters_model}. Can be converted into
data.frame with \code{\link[=as.data.frame]{as.data.frame()}} or \code{\link[data.table:as.data.table]{data.table::as.data.table()}}.}

\item{\code{.forest_data}}{More detailed data used for plotting forest.}
}}

