% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/97-regverse-package.R
\docType{package}
\name{regverse-package}
\alias{regverse}
\alias{regverse-package}
\title{regverse: regverse: Streamlined Data Modeling and Visualization in Biomedical Regression Analysis}
\description{
Provides a comprehensive suite of tools to enhance regression analysis and interpretation in the field of computational biology and clinical medicine.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WangLabCSU/regverse}
  \item \url{https://wanglabcsu.github.io/regverse/}
  \item Report bugs at \url{https://github.com/WangLabCSU/regverse/issues}
}

}
\author{
\strong{Maintainer}: Shixiang Wang \email{w_shixiang@163.com} (\href{https://orcid.org/0000-0001-9855-7357}{ORCID})

}
\keyword{internal}
