% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-run.R
\name{run}
\alias{run}
\title{Run regression model building}
\usage{
run(
  obj,
  f = c("gaussian", "coxph", "binomial", "Gamma", "inverse.gaussian", "poisson", "quasi",
    "quasibinomial", "quasipoisson"),
  exp = NULL,
  ci = 0.95,
  parallel = FALSE,
  ...
)
}
\arguments{
\item{obj}{A \code{REGObject} R6 object.}

\item{f}{A length-1 string specifying modeling function or family of \code{\link[=glm]{glm()}}, default is 'coxph'.
Other options are members of GLM family, see \code{\link[stats:family]{stats::family()}}.
'binomial' is logistic, and 'gaussian' is linear.}

\item{exp}{Logical, indicating whether or not to exponentiate the the coefficients.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).
e.g. \code{\link[survival:coxph]{survival::coxph()}}.}

\item{parallel}{If \code{TRUE}, use N-1 cores to run the task.}

\item{...}{Other parameters passing to corresponding regression model function.}
}
\value{
A \code{REGObject} R6 object.
}
\description{
Run regression model building
}
\examples{
x <- prepare(
  data = mtcars,
  vars_y = "mpg",
  vars_x = c("factor(cyl)", colnames(mtcars)[3:5]),
  vars_c = c(colnames(mtcars)[8:9], "factor(gear)")
)
x <- x |> run()
x
}
