% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-prepare.R
\name{prepare}
\alias{prepare}
\title{Prepare Data to Create a \code{REGObject}}
\usage{
prepare(data, vars_y, vars_x, vars_c)
}
\arguments{
\item{data}{A \code{data.table} storing input data.}

\item{vars_y}{Predicted variables or formulas.}

\item{vars_x}{Focal variables.}

\item{vars_c}{Covariables.}
}
\description{
Prepare Data to Create a \code{REGObject}
}
\examples{
x <- prepare(
  data = mtcars,
  vars_y = "mpg",
  vars_x = c("factor(cyl)", colnames(mtcars)[3:5]),
  vars_c = c(colnames(mtcars)[8:9], "factor(gear)")
)
x
}
