% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regverse-package.R
\docType{package}
\name{regverse-package}
\alias{regverse}
\alias{regverse-package}
\title{regverse: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://wanglabcsu.github.io/regverse/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
